chore: move primitive_datetime_serde module to utils
